requirements.txt
boto3>=1.9.188
botocore>=1.12.188
matplotlib>=3.0.3
numpy>=1.16.5
pandas>=0.24.2
pyarrow>=0.14.1
s3fs>=0.3.1
scikit-learn>=0.20.3
seaborn>=0.9.0
tqdm>=4.32.2
twine>=1.13.0
black
pytest==6.2.4
tox>=3.23.1
sklearn-pandas
faker
lightgbm
category_encoders
smart-open
sidetable